BE/fix: 임시저장

+++ Backend/src/main/java/com/ts/taesan/domain/asset/api/dto/request/TaesanPayRequest.java
@@ -12,6 +12,6 @@ public class TaesanPayRequest {
         private String shopName;
 
         @JsonProperty("pay_amt")
-        private Integer payAmt;
+        private Long payAmt;
 
     }
